# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SimpleSchedule service API.",
        "title": "SimpleSchedule API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:2811",
    "basePath": "/",
    "paths": {
...
        "/api/weeks": {
            "get": {
                "description": "Retrieve a list of all schedule weeks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weeks"
                ],
                "summary": "Get all weeks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WeeksResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new schedule week with title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weeks"
                ],
                "summary": "Create week",
                "parameters": [
                    {
                        "description": "Week creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateWeekRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        },
        "/api/weeks/{id}": {
            "get": {
                "description": "Retrieve a single week by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weeks"
                ],
                "summary": "Get week by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Week ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WeekResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the title of an existing week",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weeks"
                ],
                "summary": "Update week",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Week ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Week update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateWeekRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a week and cascade delete all its slots",
                "tags": [
                    "weeks"
                ],
                "summary": "Delete week",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Week ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        },
        "/api/weeks/{id}/schedule": {
            "get": {
                "description": "Retrieve the complete schedule for a week including all slots grouped by days",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weeks"
                ],
                "summary": "Get week schedule",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Week ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WeekScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierror.UserError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "domain.DayOfWeek": {
            "type": "string",
            "enum": [
                "monday",
                "tuesday",
                "wednesday",
                "thursday",
                "friday",
                "saturday",
                "sunday"
            ],
            "x-enum-varnames": [
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday",
                "Sunday"
            ]
        },
        "domain.SlotWithTag": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dayOfWeek": {
                    "$ref": "#/definitions/domain.DayOfWeek"
                },
                "endTime": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "startTime": {
                    "type": "integer"
                },
                "tag": {
                    "$ref": "#/definitions/domain.Tag"
                },
                "tagId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "weekId": {
                    "type": "string"
                }
            }
        },
        "domain.Tag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "hexColor": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.Week": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.WeekSchedule": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/domain.SlotWithTag"
                        }
                    }
                },
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SlotWithTag"
                    }
                },
                "week": {
                    "$ref": "#/definitions/domain.Week"
                }
            }
        },
...
        "handlers.CreateWeekRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Spring Semester 2026"
                }
            }
        },
        "handlers.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "97ec8552-9562-4d79-ab4c-1ca959663629"
                }
            }
        },
...
        "handlers.UpdateWeekRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Autumn Semester 2026"
                }
            }
        },
        "handlers.WeekResponse": {
            "type": "object",
            "properties": {
                "week": {
                    "$ref": "#/definitions/domain.Week"
                }
            }
        },
        "handlers.WeekScheduleResponse": {
            "type": "object",
            "properties": {
                "schedule": {
                    "$ref": "#/definitions/domain.WeekSchedule"
                }
            }
        },
        "handlers.WeeksResponse": {
            "type": "object",
            "properties": {
                "weeks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Week"
                    }
                }
            }
        }
    }
}